fix working process authentication
