+++ .eslintrc.json
@@ -9,46 +9,46 @@
     "plugin:react/recommended",
     "plugin:react-hooks/recommended",
     "plugin:prettier/recommended",
-		"prettier"
+    "prettier"
   ],
   "overrides": [
-		// override "simple-import-sort" config
-		{
-			"files": ["*.js", "*.jsx","*.css", "*.scss"],
-			"rules": {
-				"simple-import-sort/imports": [
-					"error",
-					{
-						"groups": [
-							// `react` first, then packages starting with a character and @
-							["^react$", "^[a-z]", "^@[a-z]"],
-							// Aliases starting with `@`
-							["^@"],
-							// Packages starting with `~`
-							["^~"],
-							// Imports starting with `../`
-							["^\\.\\.(?!/?$)", "^\\.\\./?$"],
-							// Imports starting with `./`
-							["^\\./(?=.*/)(?!/?$)", "^\\.(?!/?$)", "^\\./?$"],
-							// Style imports
-							["^.+\\.s?css$"],
-							// Side effect imports
-							["^\\u0000"]
-						]
-					}
-				]
-			}
-		}
-	],
-	"globals": {
-		"localStorage": true,
-	},
+    // override "simple-import-sort" config
+    {
+      "files": ["*.js", "*.jsx", "*.css", "*.scss"],
+      "rules": {
+        "simple-import-sort/imports": [
+          "error",
+          {
+            "groups": [
+              // `react` first, then packages starting with a character and @
+              ["^react$", "^[a-z]", "^@[a-z]"],
+              // Aliases starting with `@`
+              ["^@"],
+              // Packages starting with `~`
+              ["^~"],
+              // Imports starting with `../`
+              ["^\\.\\.(?!/?$)", "^\\.\\./?$"],
+              // Imports starting with `./`
+              ["^\\./(?=.*/)(?!/?$)", "^\\.(?!/?$)", "^\\./?$"],
+              // Style imports
+              ["^.+\\.s?css$"],
+              // Side effect imports
+              ["^\\u0000"]
+            ]
+          }
+        ]
+      }
+    }
+  ],
+  "globals": {
+    "localStorage": true
+  },
   "settings": {
     "react": {
       "version": "detect"
     }
   },
-	
+
   "parserOptions": {
     "ecmaFeatures": {
       "jsx": true
@@ -56,7 +56,7 @@
     "ecmaVersion": "latest",
     "sourceType": "module"
   },
-  "plugins": ["react","prettier", "import", "simple-import-sort"],
+  "plugins": ["react", "prettier", "import", "simple-import-sort"],
   "rules": {
     "prettier/prettier": [
       "error",
@@ -64,17 +64,17 @@
         "endOfLine": "auto"
       }
     ],
-		"no-use-before-define": ["error", { "functions": false }],
+    "no-use-before-define": ["error", { "functions": false }],
     "quotes": [1, "single"],
     "react/jsx-uses-react": "warn",
-		"react/display-name": "off",
+    "react/display-name": "off",
     "react/react-in-jsx-scope": "off",
     "react/jsx-uses-vars": "error",
     "react/jsx-fragments": "warn",
     "react/no-unescaped-entities": "warn",
-		"react/no-array-index-key": "warn",
+    "react/no-array-index-key": "warn",
     "react-hooks/rules-of-hooks": "error",
-		"simple-import-sort/imports": "warn",
-		"simple-import-sort/exports": "warn"
+    "simple-import-sort/imports": "warn",
+    "simple-import-sort/exports": "warn"
   }
 }
