+++ package.json
@@ -49,6 +49,7 @@
   "dependencies": {
     "@emotion/react": "^11.11.4",
     "@emotion/styled": "^11.11.5",
+    "@material-ui/core": "^4.12.4",
     "@mui/base": "^5.0.0-beta.40",
     "@mui/icons-material": "^5.15.18",
     "@mui/material": "^5.15.18",
