+++ src/pages/LoginPage/LoginPage.jsx
@@ -1,11 +1,42 @@
 import React, { useState } from 'react';
-import { Typography, Paper, Button, FormControl, Input, InputLabel } from '@material-ui/core';
-import { Snackbar, Alert } from '@mui/material';
-import './LoginPage.styles';
 import { signInWithEmailAndPassword } from 'firebase/auth';
-import { auth } from '../../firebase/index';
+import { Link } from 'react-router-dom';
+import { Button, FormControl, Input, InputLabel, Paper, Typography } from '@material-ui/core';
+import withStyles from '@material-ui/core/styles/withStyles';
+import { Alert, Snackbar } from '@mui/material';
 
-export const LoginPage = (props) => {
+import { auth } from '../../firebase/index';
+import { APP_PATHS } from '../../route/paths';
+
+const styles = (theme) => ({
+  main: {
+    width: 'auto',
+    display: 'block',
+    marginLeft: theme.spacing(3),
+    marginRight: theme.spacing(3),
+    [theme.breakpoints.up(400 + theme.spacing(3) * 2)]: {
+      width: 400,
+      marginLeft: 'auto',
+      marginRight: 'auto',
+    },
+  },
+  paper: {
+    marginTop: theme.spacing(8),
+    display: 'flex',
+    flexDirection: 'column',
+    alignItems: 'center',
+    padding: `${theme.spacing(2)}px ${theme.spacing(3)}px ${theme.spacing(3)}px`,
+  },
+  form: {
+    width: '100%',
+    marginTop: theme.spacing(1),
+  },
+  submit: {
+    marginTop: theme.spacing(3),
+  },
+});
+
+const LoginPage = (props) => {
   const { classes } = props;
 
   const [email, setEmail] = useState('');
@@ -83,14 +114,16 @@ export const LoginPage = (props) => {
             className={classes.submit}>
             Вход
           </Button>
-          <Button
-            type="submit"
-            fullWidth
-            variant="contained"
-            color="secondary"
-            className={classes.submit}>
-            Регистрация
-          </Button>
+          <Link to={APP_PATHS.SIGN_UP}>
+            <Button
+              type="submit"
+              fullWidth
+              variant="contained"
+              color="secondary"
+              className={classes.submit}>
+              Регистрация
+            </Button>
+          </Link>
           <Snackbar
             open={open}
             autoHideDuration={4000}
@@ -120,3 +153,5 @@ export const LoginPage = (props) => {
     </main>
   );
 };
+
+export default withStyles(styles)(LoginPage);
