+++ src/pages/SignUpPage/SignUpPage.styles.js
@@ -1,31 +1,33 @@
-import withStyles from '@material-ui/core/styles/withStyles';
+// import withStyles from '@material-ui/core/styles/withStyles';
 
-const styles = (theme) => ({
-  main: {
-    width: 'auto',
-    display: 'block',
-    marginLeft: theme.spacing(3),
-    marginRight: theme.spacing(3),
-    [theme.breakpoints.up(400 + theme.spacing(3) * 2)]: {
-      width: 400,
-      marginLeft: 'auto',
-      marginRight: 'auto',
-    },
-  },
-  paper: {
-    marginTop: theme.spacing(8),
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-    padding: `${theme.spacing(2)}px ${theme.spacing(3)}px ${theme.spacing(3)}px`,
-  },
-  form: {
-    width: '100%',
-    marginTop: theme.spacing(1),
-  },
-  submit: {
-    marginTop: theme.spacing(3),
-  },
-});
+// import { SignUpPage } from './SignUpPage';
 
-export default withStyles(styles);
+// const styles = (theme) => ({
+//   main: {
+//     width: 'auto',
+//     display: 'block',
+//     marginLeft: theme.spacing(3),
+//     marginRight: theme.spacing(3),
+//     [theme.breakpoints.up(400 + theme.spacing(3) * 2)]: {
+//       width: 400,
+//       marginLeft: 'auto',
+//       marginRight: 'auto',
+//     },
+//   },
+//   paper: {
+//     marginTop: theme.spacing(8),
+//     display: 'flex',
+//     flexDirection: 'column',
+//     alignItems: 'center',
+//     padding: `${theme.spacing(2)}px ${theme.spacing(3)}px ${theme.spacing(3)}px`,
+//   },
+//   form: {
+//     width: '100%',
+//     marginTop: theme.spacing(1),
+//   },
+//   submit: {
+//     marginTop: theme.spacing(3),
+//   },
+// });
+
+// export default withStyles(styles)(SignUpPage);
